fix: resolve max-lines lint violations from log instrumentation
Extract setupLog() to src/log/setup.mts so cli-parser.mts stays under
200 lines. Add a max-lines override (250) for http.mts which legitimately
grew with HTTP tracing instrumentation.
