+++ .oxlintrc.json
@@ -11,6 +11,12 @@
       "rules": {
         "max-lines": ["error", 500]
       }
+    },
+    {
+      "files": ["src/github/http.mts"],
+      "rules": {
+        "max-lines": ["error", 250]
+      }
     }
   ]
 }
